separate featured artist schema
